+++ 3_External_Interrupt/Inc/interrupt.h
@@ -1,9 +1,9 @@
 #ifndef INTERRUPT_H
 #define INTERRUPT_H
 
-#include "exti.h"
-#include "gpio.h"
+#include "nucleof401re.h"
 
+extern void enableInterruptIRQn(IRQn_t IRQn);
 void EXTI15_10_IRQHandler(void);
 
 #endif // INTERRUPT_H
